Some changes

+++ Middle Measurement/data_mining_utils.cpp
@@ -4,42 +4,42 @@
 
 //100 500 300 200
 
-double variance(const int* arr, int size)
+double variance(const int* arr, size_t size)
 {
     double mean = arithmetic_mean(arr, size);
     double sum = 0;
-    for (int i = 0; i < size; i++)
+    for (size_t i = 0; i < size; i++)
         sum += pow(arr[i],2);
     return (sum / size) - pow(mean, 2);
 }
 
-void calculate_quartiles(const int* arr, int size, std::vector<int>& result)
+void calculate_quartiles(const int* arr, size_t size, std::vector<int>& result)
 {
-    for (int i = 25; i <= 75; i += 25)
+    for (size_t i = 25; i <= 75; i += 25)
         result.push_back(arr[((i * size) / 100)]);
 }
 
-int range(const int* arr, int size)
+int range(const int* arr, size_t size)
 {
     return arr[max_pos(arr, size)] - arr[min_pos(arr, size)];
 }
 
 //*****************************************************
 
-double midrange(const int* arr, int size)
+double midrange(const int* arr, size_t size)
 {
     return (arr[min_pos(arr, size)] + arr[max_pos(arr, size)]) / 2.0;
 }
  
-bool mode(const int* arr, int size, std::vector<int>& modes, int& frequency)
+bool mode(const int* arr, size_t size, std::vector<int>& modes, int& frequency)
 {
     int best_frequency_value_pos = -1;
     int max_count = 0;
 
-    for (int i = 0; i < size; i++)
+    for (size_t i = 0; i < size; i++)
     {
         int count = 0;
-        for (int j = 0; j < size; j++)
+        for (size_t j = 0; j < size; j++)
             if (arr[i] == arr[j])
                 count++;
         if (count > max_count)
@@ -56,7 +56,7 @@ bool mode(const int* arr, int size, std::vector<int>& modes, int& frequency)
         for (int i = 0; i < size; i++)
         {
             int count = 0;
-            for (int j = 0; j < size; j++)
+            for (size_t j = 0; j < size; j++)
                 if (arr[i] == arr[j])
                     count++;
             if (count == max_count && 
@@ -68,7 +68,7 @@ bool mode(const int* arr, int size, std::vector<int>& modes, int& frequency)
     }
 }
 
-double median(int* arr, int size)
+double median(int* arr, size_t size)
 {
     if (size % 2 != 0)
         return arr[size / 2];
@@ -76,27 +76,27 @@ double median(int* arr, int size)
         return (arr[size / 2] + arr[size / 2 - 1]) / 2.0;
 }
 
-double trimmed_mean(int* arr, int size, double percent_ignored) {
+double trimmed_mean(int* arr, size_t size, double percent_ignored) {
     int ignored_count = (percent_ignored * size) / 100;
     int sum = 0;
-    for (int i = ignored_count; i < size - ignored_count; i++)
+    for (size_t i = ignored_count; i < size - ignored_count; i++)
         sum += arr[i];
     return (double)sum / (size - ignored_count);
 }
 
-double arithmetic_mean(const int* arr, int size) {
+double arithmetic_mean(const int* arr, size_t size) {
     int sum = 0;
-    for (int i = 0; i < size; i++)
+    for (size_t i = 0; i < size; i++)
         sum += arr[i];
     return (double)sum / size;
 }
 
 //********************************************
 
-void sort(int* arr, int size) {
-    for (int i = 0; i < size; i++) {
+void sort(int* arr, size_t size) {
+    for (size_t i = 0; i < size; i++) {
         bool is_sorted = true;
-        for (int j = 0; j < size - 1; j++) {
+        for (size_t j = 0; j < size - 1; j++) {
             if (arr[j] > arr[j + 1]) {
                 std::swap(arr[j], arr[j + 1]);
                 is_sorted = false;
@@ -107,32 +107,32 @@ void sort(int* arr, int size) {
     }
 }
 
-int min_pos(const int* arr, int size)
+int min_pos(const int* arr, size_t size)
 {
     int imin = 0;
-    for (int i = 1; i < size; i++)
+    for (size_t i = 1; i < size; i++)
         if (arr[i] < arr[imin]) imin = i;
     return imin;
 }
 
-int max_pos(const int* arr, int size)
+int max_pos(const int* arr, size_t size)
 {
     int imax = 0;
-    for (int i = 1; i < size; i++)
+    for (size_t i = 1; i < size; i++)
         if (arr[i] > arr[imax]) imax = i;
     return imax;
 }
 
-void rand_integer_data(int* arr, int size, int minrnd, int maxrnd) {
+void rand_integer_data(int* arr, size_t size, int minrnd, int maxrnd) {
     if (arr == nullptr)
         return;
-    for (int i = 0; i < size; i++)
+    for (size_t i = 0; i < size; i++)
         arr[i] = minrnd + rand() % (maxrnd - minrnd + 1);
 }
 
-void print_integer_data(const int* arr, int size) {
+void print_integer_data(const int* arr, size_t size) {
     if (arr == nullptr)
         return;
-    for (int i = 0; i < size; i++)
+    for (size_t i = 0; i < size; i++)
         std::cout << arr[i] << " ";
 }
